파일 업로드 시 기본 이미지

+++ src/main/java/GDG/whatssue/domain/file/service/impl/S3UploadService.java
@@ -1,6 +1,7 @@
 package GDG.whatssue.domain.file.service.impl;
 
 import GDG.whatssue.domain.file.service.FileUploadService;
+import GDG.whatssue.global.common.FileConst;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.CannedAccessControlList;
 import com.amazonaws.services.s3.model.ObjectMetadata;
@@ -10,6 +11,7 @@ import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
+import org.springframework.util.StringUtils;
 import org.springframework.web.multipart.MultipartFile;
 @Service
 @RequiredArgsConstructor
@@ -26,18 +28,22 @@ public class S3UploadService implements FileUploadService {
 
     @Override
     public String saveFile(MultipartFile multipartFile, String dirName) throws IOException {
-        String originalFileName = multipartFile.getOriginalFilename();
+        //저장 사진이 없으면 기본 경로 반환
+        if (multipartFile == null) {
+            return dirName + FileConst.DEFAULT_IMAGE_NAME;
+        }
 
-        String fileName = getFileName(dirName, originalFileName);
+        String originalFileName = multipartFile.getOriginalFilename();
+        String storeFileName = getFileName(dirName, originalFileName);
 
         ObjectMetadata metadata = new ObjectMetadata();
         metadata.setContentLength(multipartFile.getSize());
         metadata.setContentType(multipartFile.getContentType());
 
-        amazonS3.putObject(new PutObjectRequest(bucket, fileName, multipartFile.getInputStream(), metadata)
+        amazonS3.putObject(new PutObjectRequest(bucket, storeFileName, multipartFile.getInputStream(), metadata)
             .withCannedAcl(CannedAccessControlList.PublicRead));
 
-        return fileName;
+        return storeFileName;
     }
 
     @Override
@@ -48,7 +54,11 @@ public class S3UploadService implements FileUploadService {
 
     @Override
     public void deleteFile(String storeFileName) {
-        //TODO
+        //default 파일이면 버킷에서 삭제하지 않는다
+        if (StringUtils.substringMatch(storeFileName, 0, "default")) {
+            return;
+        }
+
         amazonS3.deleteObject(bucket, storeFileName);
     }
